requirements.txt
gunicorn
flask
firebase_admin
flask_cors
pyjwt
